Fix for #7216 (bitwise precedence). (#7217)
This is just a one-line fix for one case where the operator precedence wasn't immediately clear, and was wrong in MeshBuilder. The line in question uses bitwise AND to mask `indices[i]` with `0xFFFF` and intends to then add `offset`. Because addition binds more tightly than bitwise AND, the mask `0xFFFF` changes to some invalid mask value and only then is it applies to `indices[i]`. This commit clarifies and fixes the precedence by grouping the bitwise AND so it happens first with a valid mask, and only then will `offset` be added.