Add Vector4 class and use it where fitting (#7260)
* Add Vector4 by Antz, test.

I've completed the Vector implementation with setToRandomDirection() and isOnLine(), plus I fixed a few minor issues, but the rest of the work was by Antz.

* Fixed some typos in Vector-related docs.

This also changes several methods to use Java-style array declaration, when internally they had been using C-style array declaration. Using the C-style arrays is a style warning in IDEA and probably could be in Spotless too.

I wanted to clear up that idt() means "identical to" here, and not "identity" as it does in Quaternion, so I added one line of extra doc to each idt() in Vector3 and Vector4. I also mentioned that it compares with exact precision, not like MathUtils.isEqual() .

* Vector2.idt(Vector2)  was missing.

* Apply formatter

* Allow passing a Vector4 to a ShaderProgram.

This is very similar to how Vector3 already works, and nearly identical to how Color can be used for the same thing. A Vector4 doesn't clamp its components like Color often does, so it might be useful for different things.

* The .gwt.xml file needed updating.

* Update CHANGES.

---------

Co-authored-by: GitHub Action <action@github.com>