Fix for crashes during resizes using glfw_async
Rendering inside the resize callback when using glfw_async results in a crash `A command encoder is already encoding to this command buffer`. We need to post-pone the rendering to the next time we'd render normally. This works out most of the time. However, some flicker may occassionally occur, when control is not handed back to our rendering loop from glfwPollEvents().

Better than crashing. See also #7361